changed the model for `task` & changed the behaviour

+++ MyJira/MyJira/Controllers/TaskController.cs
@@ -9,16 +9,18 @@ using System.Web.Mvc;
 
 namespace MyJira.Controllers
 {
+    [Authorize]
     public class TaskController : Controller
     {
         private ApplicationDbContext db = new ApplicationDbContext();
         // GET: Task
+        [Authorize(Roles = "Organizer,Dev,Administrator")]
         public ActionResult Index()
         {
             var tasks = from task in db.Tasks
                         orderby task.Title
                         select task;
-            ViewBag.tasks = tasks;
+            ViewBag.tasks = tasks; 
             if (TempData.ContainsKey("message"))
             {
                 ViewBag.message = TempData["message"].ToString();
@@ -26,41 +28,111 @@ namespace MyJira.Controllers
             return View();
         }
 
+        [Authorize(Roles = "Organizer,Dev,Administrator")]
         public ActionResult Show(int id)
         {
             var task = db.Tasks.Find(id);
             ViewBag.Comments = db.Comments.Where(m => m.TaskId == id).ToList();
             ViewBag.CurrentUser = User.Identity.GetUserId();
-            return View(task);
-        }
 
-        // TODO: has to be possible only if the organizer is trying to add tasks in the curr project
-        public ActionResult New()
-        {
-            var task = new Task
+            var currentUser = db.Users.Find(User.Identity.GetUserId());
+
+            if (User.IsInRole("Dev") && currentUser.Team == null)
             {
-                ReporterId = User.Identity.GetUserId()
-            };
+                TempData["message"] = "You do not have permissions!";
+                return RedirectToAction("Index");
+            }
+
+            int? projUser = null;
+            if (User.IsInRole("Dev"))
+            {
+                projUser = currentUser.Team.ProjectId;
+            }
+
+            if (User.IsInRole("Organizer"))
+            {
+                var result =
+                    (from project in db.Projects
+                     where project.LeaderId == currentUser.Id
+                     select project.Id).ToArray();
+
+                if (result.Length == 1)
+                {
+                    projUser = result[0];
+                }
+            }
+
+            int? projTask = null;
+            if (task.Reporter.Team == null) // he is an organizer
+            {
+                var result =
+                    (from project in db.Projects
+                     where project.LeaderId == task.ReporterId
+                     select project.Id).ToArray();
+
+                if (result.Length == 1)
+                {
+                    projTask = result[0];
+                }
+            } else
+            {
+                projTask = task.Reporter.Team.ProjectId;
+            }
+
+            if (!User.IsInRole("Administrator") && (projUser == null || projTask == null || projUser != projTask))
+            {
+                TempData["message"] = "You do not have permissions!";
+                return RedirectToAction("Index");
+            }
+
+            if (TempData.ContainsKey("message"))
+            {
+                ViewBag.message = TempData["message"].ToString();
+            }
+            var project_of = db.Projects.Find(task.ProjectId);
+            ViewBag.project = project_of;
             return View(task);
         }
 
+        // TODO: has to be possible only if the organizer or team dev is trying to add tasks in the curr project
+        [Authorize(Roles = "Organizer,Administrator,Dev")]
+        public ActionResult New(int id)  // id is actually the team id from where the task is created
+        {
+            var team = db.Teams.Find(id);
+            var currentUserId = User.Identity.GetUserId();
+            var currentUser = db.Users.Find(currentUserId);
+            if ((User.IsInRole("Organizer") && team.Project.LeaderId == User.Identity.GetUserId()) || ((User.IsInRole("Dev") && (currentUser.TeamId == team.Id)) || User.IsInRole("Administrator")))
+            {
+                var task = new Task
+                {
+                    ReporterId = User.Identity.GetUserId(),
+                    TeamId = team.Id,
+                    ProjectId = team.ProjectId
+                };
+                return View(task);
+            }
+            else
+            {
+                TempData["message"] = "You do not have permissions!";
+                return RedirectToAction("Show", "Team", new { id = team.Id });
+            }
+        }
+
+        // TODO: de corectat la redirectToAction
         [HttpPost]
+        [Authorize(Roles = "Organizer,Dev,Administrator")]
         public ActionResult New(Task task)
         {
             try
             {
                 if (ModelState.IsValid)
                 {
-                    var teamId =
-                        (from user in db.Users
-                        where user.Id == task.ReporterId
-                        select user.TeamId).ToArray()[0];
-
-                // task.FinishTime = DateTime.Now; // altfel seteaza by default la o data care iese din range-ul SQL
-                db.Tasks.Add(task);
+                    // task.FinishTime = DateTime.Now; // altfel seteaza by default la o data care iese din range-ul SQL
+                    db.Tasks.Add(task);
                     db.SaveChanges();
                     TempData["message"] = "Task has been created successfully!";
-                    return RedirectToAction("Show", "Team", new { id = teamId });
+
+                    return RedirectToAction("Show", "Team", new { id = task.TeamId });
                 }
                 else
                 {
@@ -72,7 +144,7 @@ namespace MyJira.Controllers
                 return View(task);
             }
         }
-
+        [Authorize(Roles = "Organizer,Dev,Administrator")]
         public ActionResult Edit(int id)
         {
             var task = db.Tasks.Find(id);
@@ -82,6 +154,7 @@ namespace MyJira.Controllers
         }
 
         [HttpPut]
+        [Authorize(Roles = "Organizer,Dev,Administrator")]
         public ActionResult Edit(int id, Task newTask)
         {
             ViewBag.allDevs = GetAllDevs();
@@ -109,8 +182,8 @@ namespace MyJira.Controllers
 
                     var teamId =
                         (from user in db.Users
-                            where user.Id == task.ReporterId
-                            select user.TeamId).ToArray()[0];
+                         where user.Id == task.ReporterId
+                         select user.TeamId).ToArray()[0];
                     return RedirectToAction("Show", "Team", new { id = teamId });
                 }
                 else
@@ -125,6 +198,7 @@ namespace MyJira.Controllers
         }
 
         [HttpDelete]
+        [Authorize(Roles = "Organizer, Administrator")]
         public ActionResult Delete(int id)
         {
             var task = db.Tasks.Find(id);
@@ -141,7 +215,7 @@ namespace MyJira.Controllers
                        orderby user.UserName
                        select user;
 
-            return devs.Select(dev => new SelectListItem {Value = dev.Id.ToString(), Text = dev.UserName.ToString()}).ToList();
+            return devs.Select(dev => new SelectListItem { Value = dev.Id.ToString(), Text = dev.UserName.ToString() }).ToList();
         }
 
         [NonAction]
@@ -159,6 +233,7 @@ namespace MyJira.Controllers
             return selectList;
         }
 
+        [Authorize(Roles = "Organizer,Dev,Administrator")]
         public ActionResult AddComment(int id)
         {
             var comment = new Comment
@@ -172,6 +247,7 @@ namespace MyJira.Controllers
 
         [HttpPost]
         [ValidateInput(false)]
+        [Authorize(Roles = "Organizer,Dev,Administrator")]
         public ActionResult AddComment(Comment comment)
         {
             try
@@ -187,7 +263,7 @@ namespace MyJira.Controllers
                 else
                 {
                     return View(comment);
-}
+                }
             }
             catch (Exception e)
             {
@@ -195,14 +271,24 @@ namespace MyJira.Controllers
             }
         }
 
+        [Authorize(Roles = "Organizer,Dev,Administrator")]
         public ActionResult EditComment(int id)
         {
             var comment = db.Comments.Find(id);
-            return View(comment);
+            if (comment.UserId == User.Identity.GetUserId() || User.IsInRole("Administrator"))
+            {
+                return View(comment);
+            }
+            else
+            {
+                TempData["message"] = "You have not permissions !";
+                return RedirectToAction("Show/" + comment.TaskId);
+            }
         }
 
         [HttpPut]
         [ValidateInput(false)]
+        [Authorize(Roles = "Organizer,Dev,Administrator")]
         public ActionResult EditComment(int id, Comment newComment)
         {
             try
@@ -235,13 +321,20 @@ namespace MyJira.Controllers
 
 
         [HttpDelete]
+        [Authorize(Roles = "Organizer,Dev,Administrator")]
         public ActionResult DeleteComment(int id)
         {
             Comment comment = db.Comments.Find(id);
-            db.Comments.Remove(comment);
-            db.SaveChanges();
-            TempData["message"] = "Comment has been deleted successfully!";
-
+            if (comment.UserId == User.Identity.GetUserId() || User.IsInRole("Administrator"))
+            {
+                db.Comments.Remove(comment);
+                db.SaveChanges();
+                TempData["message"] = "Comment has been deleted successfully!";
+            }
+            else
+            {
+                TempData["message"] = "You have not permissions !";
+            }
             return RedirectToAction("Show/" + comment.TaskId);
         }
